requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.4
python-dotenv==1.0.0
jinja2==3.1.3
pydantic==2.5.3
httpx==0.26.0
